Remove previously supported pathological case
No longer output a file full of ones to show a difference for the last
snapshot that contains a region.  Instead, just don't output a diff
between the last snapshot that has a region and the next one that
doesn't.  In addition to making more sense and probably being better
from a semantic perspective, it also has the benefit of being far
simpler in the code and allowed removing some previously necessary
checks for this specific case.