Note that memdiff should read _regions metadata
The future is now.  Or that's what it should have been, but it isn't and
right now memdiff's approach is just to blindly diff files that have
similar region names and hope that works out totally fine.  We handle
different lengths, but we don't appropriately handle regions that change
addresses or do some of the more interesting stuff that regions may do
in some cases.  The metadata memsnap drops for each snapshot into the
_regions file could give us all the information we need to do a real and
better job of handling these cases.

This could have been done right from the beginning, but since I didn't
write it that way, here we are.

It will be a somewhat annoying refactor to do this right, but it is
worth doing to make this tool solid.  Probably should be considered a
blocker that would prevent the current code from calling itself a v1.

I guess we'll see.